requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
click==8.1.7
colorama==0.4.6
diff-match-patch==20230430
Django==5.1.1
django-debug-toolbar==4.4.6
django-import-export==4.1.1
django-redis==5.4.0
django-restframework==0.0.1
django-rq==2.10.2
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
redis==5.1.0
rq==1.16.2
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
